Move PowerBi functions into separate service
