+++ AspNetCorePowerBI/AspNetCorePowerBI/PowerBiService.cs
@@ -1,0 +1,45 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Net.Http;
+using System.Threading.Tasks;
+using AspNetCorePowerBI.Models;
+using Newtonsoft.Json.Linq;
+
+namespace AspNetCorePowerBI;
+
+public class PowerBiService
+{
+    public async Task<string> GetPowerBIAccessToken(PowerBISettings powerBISettings)
+    {
+        using var client = new HttpClient();
+
+        var form = new Dictionary<string, string>
+        {
+            ["grant_type"] = "password",
+            ["resource"] = powerBISettings.ResourceUrl,
+            ["username"] = powerBISettings.UserName,
+            ["password"] = powerBISettings.Password,
+            ["client_id"] = powerBISettings.ApplicationId.ToString(),
+            ["client_secret"] = powerBISettings.ApplicationSecret,
+            ["scope"] = "openid"
+        };
+
+        client.DefaultRequestHeaders.TryAddWithoutValidation("Content-Type", "application/x-www-form-urlencoded");
+
+        using var formContent = new FormUrlEncodedContent(form);
+
+        using var response = await client.PostAsync(powerBISettings.AuthorityUrl, formContent);
+
+        var body = await response.Content.ReadAsStringAsync();
+        var jsonBody = JObject.Parse(body);
+
+        var errorToken = jsonBody.SelectToken("error");
+
+        if (errorToken != null)
+        {
+            throw new Exception(errorToken.Value<string>());
+        }
+
+        return jsonBody.SelectToken("access_token")?.Value<string>();
+    }
+}
