+++ AspNetCorePowerBI/AspNetCorePowerBI/Program.cs
@@ -12,8 +12,8 @@ public class Program
 
     public static IHostBuilder CreateHostBuilder(string[] args)
     {
-        return Host.CreateDefaultBuilder(args)
-
+        return Host
+            .CreateDefaultBuilder(args)
             .ConfigureWebHostDefaults(webBuilder => { webBuilder.UseStartup<Startup>(); });
     }
 }
