+++ AspNetCorePowerBI/AspNetCorePowerBI/Startup.cs
@@ -1,3 +1,4 @@
+using AspNetCorePowerBI.Models;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
